refactor to reduce conflicts
