+++ app/views/decidim/initiatives/admin/initiatives/_form.html.erb
@@ -36,21 +36,33 @@
     </div>
 
     <div class="row">
-      <% unless single_initiative_type? %>
       <div class="columns xlarge-6">
-        <%= form.select :type_id,
-                        initiative_type_options,
-                        {},
+        <%= form.select :status_id,
+                        current_organization.statuses.map { |status| [status.translated_name, status.id] },
                         {
-                          disabled: !@form.signature_type_updatable?,
-                          "data-scope-selector": "initiative_decidim_scope_id",
-                          "data-scope-id": form.object.decidim_scope_id.to_s,
-                          "data-scope-search-url": decidim_initiatives.initiative_type_scopes_search_url,
-                          "data-signature-types-selector": "initiative_signature_type",
-                          "data-signature-type": current_initiative.signature_type,
-                          "data-signature-types-search-url": decidim_initiatives.initiative_type_signature_types_search_url
-                        } %>
+                          selected: current_initiative.decidim_status_id,
+                          include_blank: true
+                        }
+        %>
       </div>
+    </div>
+
+    <div class="row">
+      <% unless single_initiative_type? %>
+        <div class="columns xlarge-6">
+          <%= form.select :type_id,
+                          initiative_type_options,
+                          {},
+                          {
+                            disabled: !@form.signature_type_updatable?,
+                            "data-scope-selector": "initiative_decidim_scope_id",
+                            "data-scope-id": form.object.decidim_scope_id.to_s,
+                            "data-scope-search-url": decidim_initiatives.initiative_type_scopes_search_url,
+                            "data-signature-types-selector": "initiative_signature_type",
+                            "data-signature-type": current_initiative.signature_type,
+                            "data-signature-types-search-url": decidim_initiatives.initiative_type_signature_types_search_url
+                          } %>
+        </div>
       <% end %>
       <div class="columns xlarge-6">
         <%= form.select :decidim_scope_id, [], {}, { disabled: !@form.signature_type_updatable? } %>
@@ -86,18 +98,6 @@
                               disabled: !@form.area_updatable? %>
       </div>
     <% end %>
-
-    <div class="row">
-      <div class="columns xlarge-6">
-        <%= form.select :status_id,
-                        current_organization.statuses.map { |status| [status.translated_name, status.id] },
-                        {
-                          selected: current_initiative.decidim_status_id,
-                          include_blank: true
-                        }
-        %>
-      </div>
-    </div>
 
     <div class="row">
       <div class="columns xlarge-6">
