--- app/views/decidim/initiatives/initiatives/_filters.html.erb
@@ -1,35 +1,0 @@
-<%= initiatives_filter_form_for filter do |form| %>
-  <div class="filters__section">
-    <div class="filters__search">
-      <div class="input-group">
-        <%= form.search_field :search_text_cont, label: false, class: "input-group-field", placeholder: t(".search"), title: t(".search"), "aria-label": t(".search") %>
-        <div class="input-group-button">
-          <button type="submit" class="button">
-            <%= icon "magnifying-glass", aria_label: t(".search"), role: "img" %>
-          </button>
-        </div>
-      </div>
-    </div>
-  </div>
-
-  <%= form.check_boxes_tree :with_any_state, filter_states_values, legend_title: t(".state") %>
-
-  <%= form.check_boxes_tree :with_any_status, filter_statuses_values, legend_title: t(".status") %>
-
-  <%= form.check_boxes_tree :with_any_scope, filter_scopes_values, legend_title: t(".scope") %>
-
-  <% unless single_initiative_type? %>
-    <%= form.check_boxes_tree :with_any_type, filter_types_values, legend_title: t(".type") %>
-  <% end %>
-
-  <%= form.check_boxes_tree :with_any_area, filter_areas_values, legend_title: t(".area") %>
-
-  <% if current_user %>
-    <%= form.collection_radio_buttons :author,
-                                      [["any", t(".any")], ["myself", t(".myself")]],
-                                      :first,
-                                      :last,
-                                      legend_title: t(".author") %>
-  <% end %>
-  <%= hidden_field_tag :order, order, id: nil, class: "order_filter" %>
-<% end %>
